do not change integer type

+++ db/schema.rb
@@ -27,7 +27,7 @@ ActiveRecord::Schema[7.2].define(version: 2025_06_06_094103) do
     t.string "content_type"
     t.text "metadata"
     t.string "service_name", null: false
-    t.integer "byte_size", null: false
+    t.bigint "byte_size", null: false
     t.string "checksum"
     t.datetime "created_at", precision: nil, null: false
     t.index ["key"], name: "index_active_storage_blobs_on_key", unique: true
